% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoNormalization.R
\name{pareto_norm}
\alias{pareto_norm}
\title{PARSeC}
\usage{
pareto_norm(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object or a counting matrix with genes in rows and cellular samples in columns.}
}
\value{
If the input is a SingleCellExperiment object the function return the same object adding in the assays the Pareto normalized count matrix. If the input is a matrix pareto_norm returns a normalized matrix with the same dimensions.
}
\description{
Normalize a raw counting matrix with the shape parameter of the Pareto distribution
}
\examples{
m<-matrix(c(1,0,2,0,2,9,3,0), ncol=2)
sce<-SingleCellExperiment::SingleCellExperiment(assays=list(counts=m))

sce<-pareto_norm(sce) # SingleCellExperiment object
norm.matrix<-pareto_norm(m) # normalized matrix object

}
